Ensure we can connect to the managed VMs
In debeaac8a62b3e9d412b4f1d544b964498cc271c we started generating IP
addresses if they weren't provided in the stage inventory. It worked
great for spinning up the VMs, but unfortunately Ansible didn't know
what the IP for the host was, so couldn't connect.

To fix this we generate a temporary directory, and use that as an
inventory. We copy the original stage inventory in there, and we also
generate a file that lists all the hosts and writes a few Ansible
connection options that we require so we can SSH to the box during a
play.